dotspacemacs: remap evil-surround keybindings as an evil-surround-mode-hook
Spacemacs now lazy loads most evil packages, which unfortunately broke the
documented workaround for making 's' map to evil-substitute in visual mode, to
match vim's keybindings. So we now have to wrap their workaround in a mode hook
in order for it to actually take effect.

(Also update the link to the discussion of this in the spacemacs documentation,
whose anchor has apparently changed in the past several years...)